+++ lib/pio/open_flow.rb
@@ -14,9 +14,7 @@ module Pio
   end
 end
 
-require 'pio/open_flow/actions'
 require 'pio/open_flow/flags'
-require 'pio/open_flow/message'
 require 'pio/open_flow/open_flow_header'
 require 'pio/open_flow/phy_port'
 require 'pio/open_flow/port_number'